Fix errors with name duplication in namespaces
Libarary build without errors but examples built with errors because of
functions and variables redefinitions. Common names are global variables
and functions used for attribute interpolation. To fix that errors
common functions and variables were moved to separate file.

TODO: remove unused variables, functions and includes from textureTriangle
functions.
