+++ src_proc0/common/textureTriangle_nm0.cpp
@@ -48,28 +48,25 @@ int textureBaseLevel = 0;
 int textureMaxLevel = 1000;
 color borderColor;
 
-SECTION(".data_imu0") float initVecx [32] = {0.5,   1.5,  2.5,  3.5,  4.5,  5.5,  6.5,  7.5,  8.5,  9.5, 
-										10.5, 11.5, 12.5, 13.5, 14.5, 15.5, 16.5, 17.5, 18.5, 19.5,
-										20.5, 21.5, 22.5, 23.5, 24.5, 25.5, 26.5, 27.5, 28.5, 29.5,
-										30.5, 31.5};
-SECTION(".data_imu0") float initVecy [32] = {0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 
-										0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 
-										0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 
-										0.5, 0.5};
-SECTION(".data_imu0") float vecxf [32];
-SECTION(".data_imu0") float vecyf [32];
-SECTION(".data_imu0") float vecx [32];
-SECTION(".data_imu0") float vecy [32];
+extern float initVecx [32];
+extern float initVecy [32];
+extern float vecxf [32];
+extern float vecyf [32];
+extern float vecx [32];
+extern float vecy [32];
+extern float ones [32];
+extern float oneOverDenominator [32];
+extern float derivOneOverDenom [32];
+extern float buf0 [32];
+
+
+int minf(float a, float b);
+int maxf(float a, float b);
+void edgeFunction(float x0, float y0, float x1, float y1, float x2, float y2, float* res);
+
 SECTION(".data_imu0") float vecs [32];
 SECTION(".data_imu0") float vect [32];
 
-SECTION(".data_imu0") float ones [32] = {1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 
-										1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 
-										1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 
-										1.0, 1.0};
-
-SECTION(".data_imu0") float oneOverDenominator [32];
-SECTION(".data_imu0") float derivOneOverDenom [32];
 SECTION(".data_imu0") float vecdudx [32];
 SECTION(".data_imu0") float vecdudy [32];
 SECTION(".data_imu0") float vecdvdx [32];
@@ -95,7 +92,6 @@ SECTION(".data_imu0") float vecmu [32];
 SECTION(".data_imu0") float vecmv [32];
 SECTION(".data_imu0") float vecScaleFactor [32];
 
-SECTION(".data_imu0") float buf0 [32];
 SECTION(".data_imu0") float buf1 [32];
 SECTION(".data_imu0") float buf2 [32];
 SECTION(".data_imu0") float buf3 [32];
@@ -112,17 +108,6 @@ int min (int a, int b)
     return (b < a) ? b : a;
 }
 
-SECTION(TEXTURE_TRIANGLE_SECTION)
-int minf(float a, float b)
-{
-	return (b < a) ? b : a;
-}
-
-SECTION(TEXTURE_TRIANGLE_SECTION)
-int maxf(float a, float b)
-{
-	return (b > a) ? b : a;
-}
 
 /* Compare two floats, 1 if are equal, 0 - otherwise */
 SECTION(TEXTURE_TRIANGLE_SECTION)
@@ -1318,11 +1303,4 @@ void textureTriangle(TrianglesInfo* triangles, nm32s* pDstTriangle, int count)
     return;
 }
 
-//TEXTURING_PART
-SECTION(TEXTURE_TRIANGLE_SECTION)
-void edgeFunction(float x0, float y0, float x1, float y1, float x2, float y2, float* res)
-{
-    *res = (x2 - x0) * (y1 - y0) - (y2 - y0) * (x1 - x0);
-}
-//TEXTURING_PART
 } //end of namespace nm0_version
