+++ src_proc1/common/textureTriangle_nm1.cpp
@@ -79,6 +79,8 @@ int equalf(float a, float b)
 	}
 }
 
+void edgeFunction(float x0, float y0, float x1, float y1, float x2, float y2, float* res);
+
 SECTION(TEXTURE_TRIANGLE_SECTION)
 int getPixelValue(unsigned int x, unsigned int y, TexImage2D image, color * pixelValue)
 {
@@ -1098,13 +1100,5 @@ void textureTriangle(Pattern* patterns,
 //TEXTURING_PART
     return;
 }
-
-//TEXTURING_PART
-SECTION(TEXTURE_TRIANGLE_SECTION)
-void edgeFunction(float x0, float y0, float x1, float y1, float x2, float y2, float* res)
-{
-    *res = (x2 - x0) * (y1 - y0) - (y2 - y0) * (x1 - x0);
-}
-//TEXTURING_PART
 
 } //end of namespace nm1_version
